+++ design-toolkit/components/src/components/combobox/index.tsx
@@ -24,8 +24,12 @@ import {
   type ListBoxItemProps as AriaListBoxItemProps,
   ListBoxSection as AriaListBoxSection,
   type ListBoxSectionProps as AriaListBoxSectionProps,
+  ListLayout as AriaListLayout,
+  type ListLayoutOptions as AriaListLayoutOptions,
   Popover as AriaPopover,
   Text as AriaText,
+  Virtualizer as AriaVirtualizer,
+  type VirtualizerProps as AriaVirtualizerProps,
 } from 'react-aria-components';
 
 import { cn } from '@/lib/utils';
@@ -104,7 +108,8 @@ export interface ComboBoxProps<T extends MenuItem>
       VariantProps<typeof textFieldStyles>,
       'isDisabled' | 'isInvalid' | 'isReadOnly'
     >,
-    Omit<AriaComboBoxProps<T>, 'children'> {
+    Omit<AriaComboBoxProps<T>, 'children'>,
+    Pick<AriaVirtualizerProps<AriaListLayoutOptions>, 'layoutOptions'> {
   className?: string;
   children: ReactNode | ((item: T) => ReactNode);
   description?: string;
@@ -123,6 +128,7 @@ export function ComboBox<T extends MenuItem>({
   isReadOnly,
   label,
   placeholder,
+  layoutOptions,
   size = 'medium',
   ...props
 }: ComboBoxProps<T>) {
@@ -189,10 +195,15 @@ export function ComboBox<T extends MenuItem>({
             </AriaText>
           )}
           <AriaPopover className='w-(--trigger-width)'>
-            <AriaListBox className='grid max-h-[200px] grid-cols-[auto_1fr] overflow-y-auto overflow-x-clip rounded-medium bg-surface-overlay shadow-elevation-overlay outline outline-static-light'>
-              {/* @ts-expect-error package version mismatch TODO */}
-              {children}
-            </AriaListBox>
+            <AriaVirtualizer
+              layout={AriaListLayout}
+              layoutOptions={layoutOptions}
+            >
+              <AriaListBox className='grid max-h-[200px] grid-cols-[auto_1fr] overflow-y-auto overflow-x-clip rounded-medium bg-surface-overlay shadow-elevation-overlay outline outline-static-light'>
+                {/* @ts-expect-error package version mismatch TODO */}
+                {children}
+              </AriaListBox>
+            </AriaVirtualizer>
           </AriaPopover>
         </>
       )}